Response Class Updated to models
