+++ example/lib/main.dart
@@ -1,8 +1,9 @@
 import 'dart:async';
 import 'dart:convert';
 
-import 'package:cell_info/CellResponse.dart';
+import 'package:cell_info/cellResponse.dart';
 import 'package:cell_info/cell_info.dart';
+import 'package:cell_info/models/common/cell_type.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
@@ -36,7 +37,7 @@ class _MyAppState extends State<MyApp> {
 
       cellsResponse = CellsResponse.fromJson(body);
 
-      CellType currentCellInFirstChip = cellsResponse.primaryCellList[1];
+      CellType currentCellInFirstChip = cellsResponse.primaryCellList[0];
       if (currentCellInFirstChip.type == "LTE") {
         currentDBM =
             "LTE dbm = " + currentCellInFirstChip.lte.signalLTE.dbm.toString();
