+++ lib/models/lte/lte_parameters.dart
@@ -1,63 +1,76 @@
-import '../common/cell_info.dart';
 
-class LTEParameters extends BasicCellInfo {
+import 'package:cell_info/models/lte/lte_band.dart';
+import 'package:cell_info/models/lte/lte_signal.dart';
+
+import '../common/network.dart';
+
+class Lte {
+  BandLTE bandLTE;
+  int bandwidth;
+  int cid;
   int eci;
   int enb;
-  int cid;
-  int tac;
   int pci;
-  int bandWidth;
-  int rssi;
-  double rsrp;
-  double rsrq;
-  int cqi;
-  int timingAdvance;
-  double snr;
+  SignalLTE signalLTE;
+  int tac;
+  String connectionStatus;
+  Network network;
+  int subscriptionId;
+  String type;
 
-  LTEParameters({
-    this.eci,
-    this.enb,
-    this.cid,
-    this.tac,
-    this.pci,
-    this.bandWidth,
-    this.rssi,
-    this.rsrp,
-    this.rsrq,
-    this.cqi,
-    this.timingAdvance,
-    this.snr,
-  });
+  Lte(
+      {this.bandLTE,
+        this.bandwidth,
+        this.cid,
+        this.eci,
+        this.enb,
+        this.pci,
+        this.signalLTE,
+        this.tac,
+        this.connectionStatus,
+        this.network,
+        this.subscriptionId,
+        this.type});
 
-  LTEParameters.fromJson(Map<String, dynamic> map) {
-    this.eci = map['eci'];
-    this.enb = map['enb'];
-    this.cid = map['cid'];
-    this.tac = map['tac'];
-    this.pci = map['pci'];
-    this.bandWidth = map['bandWidth'];
-    this.rssi = map['rssi'];
-    this.rsrp = map['rsrp'];
-    this.rsrq = map['rsrq'];
-    this.cqi = map['cqi'];
-    this.timingAdvance = map['timingAdvance'];
-    this.snr = map['snr'];
+  Lte.fromJson(Map<String, dynamic> json) {
+    bandLTE =
+    json['bandLTE'] != null ? new BandLTE.fromJson(json['bandLTE']) : null;
+    bandwidth = json['bandwidth'];
+    cid = json['cid'];
+    eci = json['eci'];
+    enb = json['enb'];
+    pci = json['pci'];
+    signalLTE = json['signalLTE'] != null
+        ? new SignalLTE.fromJson(json['signalLTE'])
+        : null;
+    tac = json['tac'];
+    connectionStatus = json['connectionStatus'];
+    network =
+    json['network'] != null ? new Network.fromJson(json['network']) : null;
+    subscriptionId = json['subscriptionId'];
+    type = json['type'];
+  }
 
-    Map<String, dynamic> toJson() {
-      final Map<String, dynamic> data = new Map<String, dynamic>();
-      data['eci'] = this.eci;
-      data['enb'] = this.enb;
-      data['cid'] = this.cid;
-      data['tac'] = this.tac;
-      data['pci'] = this.pci;
-      data['bandWidth'] = this.bandWidth;
-      data['rssi'] = this.rssi;
-      data['rsrp'] = this.rsrp;
-      data['rsrq'] = this.rsrq;
-      data['cqi'] = this.cqi;
-      data['timingAdvance'] = this.timingAdvance;
-      data['snr'] = this.snr;
-      return data;
+  Map<String, dynamic> toJson() {
+    final Map<String, dynamic> data = new Map<String, dynamic>();
+    if (this.bandLTE != null) {
+      data['bandLTE'] = this.bandLTE.toJson();
     }
+    data['bandwidth'] = this.bandwidth;
+    data['cid'] = this.cid;
+    data['eci'] = this.eci;
+    data['enb'] = this.enb;
+    data['pci'] = this.pci;
+    if (this.signalLTE != null) {
+      data['signalLTE'] = this.signalLTE.toJson();
+    }
+    data['tac'] = this.tac;
+    data['connectionStatus'] = this.connectionStatus;
+    if (this.network != null) {
+      data['network'] = this.network.toJson();
+    }
+    data['subscriptionId'] = this.subscriptionId;
+    data['type'] = this.type;
+    return data;
   }
 }
