+++ lib/models/wcdma/wcdma_parameters.dart
@@ -1,27 +1,47 @@
-import 'package:cell_info/models/common/common_wtdscdma_parameters.dart';
+import '../common/network.dart';
+import 'cdma_band.dart';
+import 'cdma_signal.dart';
 
-class WCDMAParameters extends CommonWCDMAParameters {
-  int psc;
-  int ecio;
-  int ecno;
+class Wcdma {
+  BandWCDMA bandWCDMA;
+  SignalWCDMA signalWCDMA;
+  String connectionStatus;
+  Network network;
+  String type;
 
-  WCDMAParameters({
-    this.psc,
-    this.ecio,
-    this.ecno,
-  });
+  Wcdma(
+      {this.bandWCDMA,
+        this.signalWCDMA,
+        this.connectionStatus,
+        this.network,
+        this.type});
 
-  WCDMAParameters.fromJson(Map<String, dynamic> map) {
-    this.psc = map['psc'];
-    this.ecio = map['ecio'];
-    this.ecno = map['ecno'];
+  Wcdma.fromJson(Map<String, dynamic> json) {
+    bandWCDMA = json['bandWCDMA'] != null
+        ? new BandWCDMA.fromJson(json['bandWCDMA'])
+        : null;
+    signalWCDMA = json['signalWCDMA'] != null
+        ? new SignalWCDMA.fromJson(json['signalWCDMA'])
+        : null;
+    connectionStatus = json['connectionStatus'];
+    network =
+    json['network'] != null ? new Network.fromJson(json['network']) : null;
+    type = json['type'];
+  }
 
-    Map<String, dynamic> toJson() {
-      final Map<String, dynamic> data = new Map<String, dynamic>();
-      data['psc'] = this.psc;
-      data['ecio'] = this.ecio;
-      data['ecno'] = this.ecno;
-      return data;
+  Map<String, dynamic> toJson() {
+    final Map<String, dynamic> data = new Map<String, dynamic>();
+    if (this.bandWCDMA != null) {
+      data['bandWCDMA'] = this.bandWCDMA.toJson();
     }
+    if (this.signalWCDMA != null) {
+      data['signalWCDMA'] = this.signalWCDMA.toJson();
+    }
+    data['connectionStatus'] = this.connectionStatus;
+    if (this.network != null) {
+      data['network'] = this.network.toJson();
+    }
+    data['type'] = this.type;
+    return data;
   }
 }
